Finish doc service (#34)
* adjust upload doc

* add reponse & doc list

* add redis

* add arq

* fix cli

* finish doc upload

* close ci test

* fix extract

+++ app/models/document.py
@@ -4,17 +4,15 @@ from datetime import datetime
 from typing import TYPE_CHECKING
 
 import aiofiles
-from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..config import settings
 from ..database import Base
 from ..schemas.document import DocState, FileType
+from ..settings import settings
 from .keyword import document_keywords
 
 if TYPE_CHECKING:
     from .keyword import Keyword
-    from .subject import Subject
 
 
 class Document(Base):
@@ -22,17 +20,14 @@ class Document(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     title: Mapped[str] = mapped_column(nullable=False)
-    file_name: Mapped[str] = mapped_column(nullable=False)
+    local_file_name: Mapped[str] = mapped_column(nullable=False)
     file_type: Mapped[FileType] = mapped_column(nullable=False)
-    subject_id: Mapped[int] = mapped_column(ForeignKey("subjects.id"))
     state: Mapped[DocState] = mapped_column(default=DocState.UPLOADED, nullable=False)
     word_count: Mapped[int | None] = mapped_column(default=None)
     created_at: Mapped[datetime] = mapped_column(default=datetime.now, nullable=False)
     updated_at: Mapped[datetime] = mapped_column(
         default=datetime.now, onupdate=datetime.now, nullable=False
     )
-
-    subject: Mapped[Subject] = relationship(back_populates="documents")
     keywords: Mapped[set[Keyword]] = relationship(
         "Keyword",
         secondary=document_keywords,
@@ -42,19 +37,34 @@ class Document(Base):
     )
 
     @property
+    def file_name(self):
+        """获取文件名"""
+        return f"{self.title}.{self.file_type}"
+
+    @property
+    def file_size(self):
+        """获取文件大小"""
+        return self.upload_path.stat().st_size
+
+    @property
     def upload_path(self):
         """获取原始上传文件路径"""
-        return settings.UPLOAD_DIR / f"{self.file_name}.{self.file_type}"
+        return settings.UPLOAD_DIR / f"{self.local_file_name}.{self.file_type}"
 
     @property
     def extracted_path(self):
         """获取提取文本的文件路径"""
-        return settings.RAW_TEXT_DIR / f"{self.file_name}.txt"
+        return settings.RAW_TEXT_DIR / f"{self.local_file_name}.txt"
 
     @property
     def normalized_path(self):
         """获取标准化文本的文件路径"""
-        return settings.NORM_TEXT_DIR / f"{self.file_name}.txt"
+        return settings.NORM_TEXT_DIR / f"{self.local_file_name}.txt"
+
+    @property
+    def url(self):
+        """获取文档的下载URL"""
+        return f"{settings.API_V1_STR}/documents/{self.id}/file"
 
     def get_path(self, state: DocState):
         """根据处理阶段获取对应的文件路径"""
@@ -67,12 +77,16 @@ class Document(Base):
     def create_dirs(self):
         """创建文档所需的所有目录"""
         for state in DocState:
+            if not state.is_finished:
+                continue
             file_path = self.get_path(state)
             file_path.parent.mkdir(parents=True, exist_ok=True)
 
     def delete_dirs(self):
         """删除文档所需的所有目录"""
         for state in DocState:
+            if not state.is_finished:
+                continue
             file_path = self.get_path(state)
             file_path.unlink(missing_ok=True)
